requirements.txt
pillow
opencv-python
google-cloud-vision
langchain-community
langchain-openai
flask
faiss-cpu==1.7.4
openai
python-dotenv
document
pytesseract
langchain-google-genai